requirements.txt
apache-airflow==2.7.2
pyspark==3.5.1
python-dotenv==1.0.0
